requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
python-multipart==0.0.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==0.19.0
SQLAlchemy==1.4.23
pymysql==1.0.2
mysql-connector-python==8.0.26
redis==3.5.3
aiofiles==0.7.0
prometheus-client==0.11.0
APScheduler==3.8.1
python-json-logger==2.0.2
requests==2.26.0
httpx==0.19.0 